refactor: custom hook 분리
