+++ src/hooks/useLoading.ts
@@ -1,0 +1,11 @@
+import { useState } from 'react';
+
+export const useLoading = () => {
+  const [isLoading, setIsLoading] = useState(false);
+
+  const setLoading = (loading: boolean) => {
+    setIsLoading(loading);
+  };
+
+  return { isLoading, setLoading };
+};
